apps/brain/requirements.txt
httpx>=0.28.1
python-dotenv>=1.0.1
openai>=1.61.0

